refactor: Update longitude calculation in _buildMarkerWidget
This commit updates the _buildMarkerWidget method in the map.dart file to improve the calculation of the longitude value for the vessel marker. It ensures that the longitude value is within the range of -180 to 180 degrees by applying appropriate adjustments. This enhancement improves the accuracy of the vessel marker positions on the map.